add ProgressBar with new Toolbar to show loading
A ProgressBar is displayed when a page is rendering; the ProgressBar is
hidden after the page has been displayed to the screen. In order to make
the ProgressBar show under the ActionBar, we use our own Toolbar
integrated into the Activity's layout instead of the one provided by the
system. (This also gives us a bit more control over styling.)

The ProgressBar is set to indeterminate. Zooming in does not show the
ProgressBar.