move initial showSystemUi call to loadPdf
The reasoning for calling showSystemUi() in the onCreate method
before was to make the transitions for immersive mode more smooth.
However, it's only needed when the PDF has actually loaded, i.e. when
the immersive mode makes sense to use.

Before, Snackbars were partially covered by the navigation UI due to the
showSystemUi() setting the View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
flag. Moving it so that the function is only called when a PDF is loaded
fixes this issue.