add redundant setBlockNetworkLoads(true)
This defaults to true when the app doesn't have the INTERNET permission
but it makes sense to explicitly disable it in case the app was extended
to allow the INTERNET permission. We would still intercept all of the
requests to manually handle them regardless, but it's good to have this
enabled anyway.